# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalResponse": {
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "transfer",
              "withdrawal"
            ],
            "title": "Kind",
            "type": "string"
          },
          "movement_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movement Id"
          },
          "request": {
            "$ref": "#/components/schemas/RequestedMovement"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected",
              "expired",
              "executed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agent_id",
          "kind",
          "status",
          "request",
          "movement_id",
          "failure_code",
          "expires_at",
          "decided_at",
          "created_at"
        ],
        "title": "ApprovalResponse",
        "type": "object"
      },
...
      "AwaitingApprovalResponse": {
        "description": "What an agent is answered when what it asked for waits for its owner: nothing has\nmoved, and this is the request the owner will see.",
        "properties": {
          "approval_request": {
            "$ref": "#/components/schemas/ApprovalResponse"
          }
        },
        "required": [
          "approval_request"
        ],
        "title": "AwaitingApprovalResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RequestedMovement": {
        "description": "What the agent asked for. A transfer names a recipient; a withdrawal names a\nbeneficiary or an address.",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "beneficiary_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "recipient_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Id"
          },
          "to_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Address"
          }
        },
        "required": [
          "asset",
          "amount"
        ],
        "title": "RequestedMovement",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WithdrawalPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WithdrawalResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "WithdrawalPageResponse",
        "type": "object"
      },
      "WithdrawalRequest": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "maxLength": 320,
            "title": "Amount",
            "type": "string"
          },
          "asset": {
            "maxLength": 320,
            "title": "Asset",
            "type": "string"
          },
          "beneficiary_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "to_address": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Address"
          }
        },
        "required": [
          "asset",
          "amount"
        ],
        "title": "WithdrawalRequest",
        "type": "object"
      },
      "WithdrawalResponse": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "beneficiary_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "fee": {
            "title": "Fee",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "bank",
              "chain"
            ],
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "enum": [
              "held",
              "submitting",
              "submitted",
              "completed",
              "failed",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          },
          "to_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Address"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "asset",
          "amount",
          "fee",
          "kind",
          "beneficiary_id",
          "to_address",
          "failure_reason",
          "created_at",
          "updated_at"
        ],
        "title": "WithdrawalResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "An access token from /v1/auth/login, or an agent API key.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Corridor",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/withdrawals": {
      "get": {
        "operationId": "list_withdrawals_v1_withdrawals_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The withdrawals requested, newest first",
        "tags": [
          "withdrawals"
        ]
      },
      "post": {
        "operationId": "request_withdrawal_v1_withdrawals_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponse",
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WithdrawalResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AwaitingApprovalResponse"
                    }
                  ],
                  "title": "Response 202 Request Withdrawal V1 Withdrawals Post"
                }
              }
            },
            "description": "The withdrawal, with its funds reserved. Or, for an agent that asked for more than its owner lets it send unseen, the request the owner has been asked to approve: nothing is reserved."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw to a saved bank account or to an address",
        "tags": [
          "withdrawals"
        ]
      }
    },
    "/v1/withdrawals/{withdrawal_id}": {
      "get": {
        "operationId": "get_withdrawal_v1_withdrawals__withdrawal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "withdrawal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Withdrawal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One withdrawal",
        "tags": [
          "withdrawals"
        ]
      }
    },
    "/v1/withdrawals/{withdrawal_id}/cancel": {
      "post": {
        "operationId": "cancel_withdrawal_v1_withdrawals__withdrawal_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "withdrawal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Withdrawal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Call back a withdrawal that has not been sent",
        "tags": [
          "withdrawals"
        ]
      }
    }
  }
}